+++ src/test/index.js
@@ -35,8 +35,11 @@ import testFeeds from './testFeeds.js';
 import testFloodgauges from './testFloodgauges.js';
 import testInfrastructure from './testInfrastructure.js';
 import testFloods from './testFloods.js';
+import testFloodsArchive from './testFloodsArchive';
+import testFloodsTimeseries from './testFloodsTimeseries';
 import testReports from './testReports.js';
 import testReportsArchive from './testReportsArchive';
+import testReportsTimeseries from './testReportsTimeseries';
 import testCAP from './testCAP.js';
 import testDB from './testDB.js';
 
@@ -76,9 +79,12 @@ let token = jwt.sign({},
     testFloodgauges(app);
     testInfrastructure(app);
     testFloods(app, token);
+    testFloodsArchive(app);
+    testFloodsTimeseries(app);
     testReports(app, reportid, createdAt);
-    testReportsArchive(app, createdAt);
-    testCAP(logger);
+    testReportsArchive(app);
+    testReportsTimeseries(app);
+    testCAP(config, logger);
     testDB();
 
     // Removes dummy data