+++ src/test/testFloodsArchive.js
@@ -1,8 +1,8 @@
 /* eslint-disable max-len */
 /**
  * testReportsArchive module
- * @module test/testReportsArchive
- * A module to test the /reports/archive endpoint
+ * @module test/testFloodsArchive
+ * A module to test the /floods/archive endpoint
  */
 
 import * as test from 'unit.js';
@@ -12,89 +12,23 @@ import * as test from 'unit.js';
 // - Entered through cards.js (or add a new call to cards here)
 
 /**
- * Test reports archive endpoint
- * @function testReportsArchive
+ * Test floods archive endpoint
+ * @function testFloodsArchive
  * @param {Object} app - CogniCity server app object
  */
 export default function(app) {
   // Reports endpoint
   describe('Reports Archive Endpoint', function() {
     // Can get reports between given timestamps
-    let end = new Date('2017-06-07T00:00:00+0700');
-    end.setHours(end.getHours() + 72);
-    end = end.toISOString().slice(0, -5)+'Z';
-
-    it('Can get reports between given timestamps', function(done) {
+    it('Can get floods between given timestamps', function(done) {
         test.httpAgent(app)
-          .get('/reports/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-08T23:00:00%2B0700')
+          .get('/floods/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-08T23:00:00%2B0700')
           .expect(200)
           .expect('Content-Type', /json/)
           .end(function(err, res) {
             if (err) {
               test.fail(err.message + ' ' + JSON.stringify(res));
             } else {
-              done();
-            }
-         });
-      });
-
-    it('Can get reports in given city', function(done) {
-        test.httpAgent(app)
-          .get('/reports/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-08T23:00:00%2B0700&city=jbd')
-          .expect(200)
-          .expect('Content-Type', /json/)
-          .end(function(err, res) {
-            if (err) {
-              test.fail(err.message + ' ' + JSON.stringify(res));
-            } else {
-              done();
-            }
-         });
-      });
-
-    it('Catches bad city name', function(done) {
-        test.httpAgent(app)
-          .get('/reports/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-08T23:00:00%2B0700&city=123')
-          .expect(400)
-          .expect('Content-Type', /json/)
-          .end(function(err, res) {
-            if (err) {
-              test.fail(err.message + ' ' + JSON.stringify(res));
-            } else {
-              done();
-            }
-         });
-      });
-
-    it('Can get reports between given timestamps as geojson', function(done) {
-        test.httpAgent(app)
-          .get('/reports/archive?start=2017-06-07T00:00:00%2B0700&end='+end+'&format=json&geoformat=geojson')
-          .expect(200)
-          .expect('Content-Type', /json/)
-          .end(function(err, res) {
-            if (err) {
-              test.fail(err.message + ' ' + JSON.stringify(res));
-            } else {
-              test.value(res.body.result.type).is('FeatureCollection');
-              test.value(res.body.result.features[0].properties.status)
-                .is('confirmed');
-              test.value(res.body.result.features[0].properties.disaster_type)
-                .is('flood');
-              done();
-            }
-         });
-      });
-
-    it('Can get reports between timestamps as topojson', function(done) {
-        test.httpAgent(app)
-          .get('/reports/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-08T23:00:00%2B0700&format=json&geoformat=topojson')
-          .expect(200)
-          .expect('Content-Type', /json/)
-          .end(function(err, res) {
-            if (err) {
-              test.fail(err.message + ' ' + JSON.stringify(res));
-            } else {
-              test.value(res.body.result.type).is('Topology');
               done();
             }
          });
@@ -103,7 +37,7 @@ export default function(app) {
     // Can catch no start parameter
     it('Required start parameter by default', function(done) {
         test.httpAgent(app)
-          .get('/reports/archive?end=2017-02-22T07:00:00%2B0700')
+          .get('/floods/archive?end=2017-02-22T07:00:00%2B0700')
           .expect(400)
           .expect('Content-Type', /json/)
           .end(function(err, res) {
@@ -118,7 +52,7 @@ export default function(app) {
     // Can catch no end parameter
     it('Required end parameter by default', function(done) {
         test.httpAgent(app)
-          .get('/reports/archive?start=2017-02-22T07:00:00%2B0700')
+          .get('/floods/archive?start=2017-02-22T07:00:00%2B0700')
           .expect(400)
           .expect('Content-Type', /json/)
           .end(function(err, res) {
@@ -174,7 +108,8 @@ export default function(app) {
                 }
              });
           });
-          // Catches start - end time window greater than one week
+
+        // Catches start - end time window greater than one week
         it('Catch large time windows', function(done) {
             test.httpAgent(app)
               .get('/floods/archive?start=2017-06-07T00:00:00%2B0700&end=2017-06-15T23:00:00%2B0700')